去除没有用的标识符或Import

+++ db/org.tinygroup.tinydb/src/main/java/org/tinygroup/tinydb/dialect/impl/AbstractDialect.java
@@ -81,7 +81,7 @@ public abstract class AbstractDialect implements Dialect {
 	}
 	
 	public int nextIntValue() throws DataAccessException {
-		return (int) getNextKey();
+		return getNextKey();
 	}
 
 	public long nextLongValue() throws DataAccessException {
